Fix tests

+++ database/migrations/admin/2023_01_10_000400_add_delivery_comment_orders_table.php
@@ -6,8 +6,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up()
     {
         if (Schema::hasColumn('orders', 'delivery_comment')) {
@@ -21,6 +20,10 @@ return new class extends Migration
 
     public function down()
     {
+        if (!Schema::hasColumn('orders', 'delivery_comment')) {
+            return;
+        }
+
         Schema::table('orders', function (Blueprint $table) {
             $table->dropColumn('delivery_comment');
         });
